+++ helpers/compile/configs.ts
@@ -1,25 +1,40 @@
+import path from 'node:path'
+
 import { BuildOptions } from './build'
 
-export const adapterConfig: BuildOptions[] = [
-  {
-    name: 'cjs',
-    format: 'cjs',
-    bundle: true,
-    entryPoints: ['src/index.ts'],
-    outfile: 'dist/index',
-    outExtension: { '.js': '.js' },
-    emitTypes: true,
-  },
-  {
-    name: 'esm',
-    format: 'esm',
-    bundle: true,
-    entryPoints: ['src/index.ts'],
-    outfile: 'dist/index',
-    outExtension: { '.js': '.mjs' },
-    emitTypes: true,
-  },
-]
+type AdapterEntry = {
+  entry: string
+  outfile: string
+}
+
+export function createAdapterConfig(entries: AdapterEntry[]): BuildOptions[] {
+  return entries.flatMap(({ entry, outfile }) => {
+    const baseName = path.basename(outfile)
+
+    return [
+      {
+        name: `${baseName}-cjs`,
+        format: 'cjs' as const,
+        bundle: true,
+        entryPoints: [entry],
+        outfile,
+        outExtension: { '.js': '.js' },
+        emitTypes: true,
+      },
+      {
+        name: `${baseName}-esm`,
+        format: 'esm' as const,
+        bundle: true,
+        entryPoints: [entry],
+        outfile,
+        outExtension: { '.js': '.mjs' },
+        emitTypes: true,
+      },
+    ]
+  })
+}
+
+export const adapterConfig = createAdapterConfig([{ entry: 'src/index.ts', outfile: 'dist/index' }])
 
 export const unbundledConfig: BuildOptions[] = [
   {
