+++ packages/adapter-d1/package.json
@@ -2,18 +2,31 @@
   "name": "@prisma/adapter-d1",
   "version": "0.0.0",
   "description": "Prisma's driver adapter for Cloudflare D1",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
+  "main": "dist/index-node.js",
+  "module": "dist/index-node.mjs",
+  "types": "dist/index-node.d.ts",
   "exports": {
     ".": {
-      "require": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
+      "types": "./dist/index-node.d.ts",
+      "node": {
+        "types": "./dist/index-node.d.ts",
+        "require": "./dist/index-node.js",
+        "import": "./dist/index-node.mjs"
       },
-      "import": {
-        "types": "./dist/index.d.mts",
-        "default": "./dist/index.mjs"
+      "workerd": {
+        "types": "./dist/index-workerd.d.ts",
+        "require": "./dist/index-workerd.js",
+        "import": "./dist/index-workerd.mjs"
+      },
+      "browser": {
+        "types": "./dist/index-workerd.d.ts",
+        "require": "./dist/index-workerd.js",
+        "import": "./dist/index-workerd.mjs"
+      },
+      "default": {
+        "types": "./dist/index-node.d.ts",
+        "require": "./dist/index-node.js",
+        "import": "./dist/index-node.mjs"
       }
     }
   },
