+++ packages/adapter-d1/src/d1.ts
@@ -1,6 +1,3 @@
-import fs from 'node:fs'
-import path from 'node:path'
-
 import { D1Database } from '@cloudflare/workers-types'
 import { SqlDriverAdapter, SqlDriverAdapterFactory } from '@prisma/driver-adapter-utils'
 
@@ -35,34 +32,4 @@ export class PrismaD1<Args extends D1Database | D1HttpParams = D1Database> imple
 
 type PrismaD1Interface<Params> = SqlDriverAdapterFactory & {
   connectToShadowDb: Params extends D1HttpParams ? () => Promise<SqlDriverAdapter> : undefined
-}
-
-const localD1DatabasePath = path.join('.wrangler', 'state', 'v3', 'd1', 'miniflare-D1DatabaseObject')
-
-/**
- * Lists all local D1 databases found in the Wrangler state directory.
- *
- * This function scans the `.wrangler/state/v3/d1/miniflare-D1DatabaseObject` directory
- * for SQLite database files and returns their paths.
- *
- * @returns An array of file paths to local D1 database files
- *
- * @example
- * ```typescript
- * import { defineConfig } from '@prisma/config'
- * import { listLocalDatabases } from '@prisma/adapter-d1'
- *
- * export default defineConfig({
- *   engine: 'classic',
- *   datasource: {
- *     url: `file:${listLocalDatabases().pop()}`,
- *   },
- * })
- * ```
- */
-export function listLocalDatabases(): string[] {
-  const cwd = process.cwd()
-  const d1DirPath = path.join(cwd, localD1DatabasePath)
-  const files = fs.readdirSync(d1DirPath)
-  return files.filter((file) => file.endsWith('.sqlite')).map((file) => path.join(d1DirPath, file))
 }
