+++ packages/adapter-d1/src/index-node.ts
@@ -1,0 +1,2 @@
+export * from './index-workerd'
+export { listLocalDatabases } from './node/list-local-databases'
