+++ packages/adapter-d1/src/index-workerd.ts
@@ -1,3 +1,3 @@
-export { listLocalDatabases, PrismaD1 } from './d1'
+export { PrismaD1 } from './d1'
 // exported for backwards compatibility
 export { PrismaD1HttpAdapterFactory as PrismaD1Http } from './d1-http'
