+++ tsconfig.build.bundle.json
@@ -23,7 +23,7 @@
       "@prisma/engines": ["./packages/engines/src"],
       "@prisma/instrumentation": ["./packages/instrumentation/src"],
       "@prisma/driver-adapter-utils": ["./packages/driver-adapter-utils/src"],
-      "@prisma/adapter-d1": ["./packages/adapter-d1/src"],
+      "@prisma/adapter-d1": ["./packages/adapter-d1/src/index-node.ts"],
       "@prisma/adapter-libsql": ["./packages/adapter-libsql/src"],
       "@prisma/adapter-neon": ["./packages/adapter-neon/src"],
       "@prisma/adapter-pg": ["./packages/adapter-pg/src"],
